+++ src/main/java/ca/corefacility/bioinformatics/irida/model/workflow/submission/AnalysisSubmission.java
@@ -10,6 +10,7 @@ import java.util.Objects;
 import java.util.Optional;
 import java.util.Set;
 import java.util.UUID;
+import java.util.stream.Collectors;
 
 import javax.persistence.CascadeType;
 import javax.persistence.CollectionTable;
@@ -50,7 +51,6 @@ import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Maps;
-import com.google.common.collect.Sets;
 
 import ca.corefacility.bioinformatics.irida.exceptions.AnalysisAlreadySetException;
 import ca.corefacility.bioinformatics.irida.model.IridaResourceSupport;
@@ -59,6 +59,7 @@ import ca.corefacility.bioinformatics.irida.model.enums.AnalysisCleanedState;
 import ca.corefacility.bioinformatics.irida.model.enums.AnalysisState;
 import ca.corefacility.bioinformatics.irida.model.project.ReferenceFile;
 import ca.corefacility.bioinformatics.irida.model.sequenceFile.SequenceFilePair;
+import ca.corefacility.bioinformatics.irida.model.sequenceFile.SequencingObject;
 import ca.corefacility.bioinformatics.irida.model.sequenceFile.SingleEndSequenceFile;
 import ca.corefacility.bioinformatics.irida.model.user.User;
 import ca.corefacility.bioinformatics.irida.model.workflow.analysis.Analysis;
@@ -118,12 +119,8 @@ public class AnalysisSubmission extends IridaResourceSupport implements MutableI
 	private String remoteWorkflowId;
 	
 	@ManyToMany(fetch = FetchType.EAGER, cascade = CascadeType.DETACH)
-	@JoinTable(name = "analysis_submission_sequence_file_single_end", joinColumns = @JoinColumn(name = "analysis_submission_id", nullable = false), inverseJoinColumns = @JoinColumn(name = "sequencing_object_id", nullable = false))
-	private Set<SingleEndSequenceFile> inputFilesSingleEnd;
-
-	@ManyToMany(fetch = FetchType.EAGER, cascade = CascadeType.DETACH)
-	@JoinTable(name = "analysis_submission_sequence_file_pair", joinColumns = @JoinColumn(name = "analysis_submission_id", nullable = false), inverseJoinColumns = @JoinColumn(name = "sequence_file_pair_id", nullable = false))
-	private Set<SequenceFilePair> inputFilesPaired;
+	@JoinTable(name = "analysis_submission_sequencing_object", joinColumns = @JoinColumn(name = "analysis_submission_id", nullable = false), inverseJoinColumns = @JoinColumn(name = "sequencing_object_id", nullable = false))
+	private Set<SequencingObject> inputFiles;
 
 	@ElementCollection(fetch = FetchType.EAGER)
 	@MapKeyColumn(name = "name", nullable = false)
@@ -193,12 +190,11 @@ public class AnalysisSubmission extends IridaResourceSupport implements MutableI
 		this();
 		checkNotNull(builder.workflowId, "workflowId is null");
 
-		checkArgument(builder.inputFilesSingleEnd != null || builder.inputFilesPaired != null,
-				"all input file collections are null.  You must supply at least one set of input files");
+		checkArgument(builder.inputFiles != null,
+				"input file collection is null.  You must supply at least one set of input files");
 
 		this.name = (builder.name != null) ? builder.name : "Unknown";
-		this.inputFilesSingleEnd = (builder.inputFilesSingleEnd != null) ? builder.inputFilesSingleEnd : Sets.newHashSet();
-		this.inputFilesPaired = (builder.inputFilesPaired != null) ? builder.inputFilesPaired : Sets.newHashSet();
+		this.inputFiles = builder.inputFiles;
 		this.inputParameters = (builder.inputParameters != null) ? ImmutableMap.copyOf(builder.inputParameters)
 				: ImmutableMap.of();
 		this.referenceFile = builder.referenceFile;
@@ -244,7 +240,11 @@ public class AnalysisSubmission extends IridaResourceSupport implements MutableI
 	 */	
 	@JsonIgnore
 	public Set<SingleEndSequenceFile> getInputFilesSingleEnd() {
-		return inputFilesSingleEnd;
+		return inputFiles.stream().filter(f -> {
+			return f instanceof SingleEndSequenceFile;
+		}).map(f -> {
+			return (SingleEndSequenceFile) f;
+		}).collect(Collectors.toSet());
 	}
 
 	/**
@@ -254,7 +254,11 @@ public class AnalysisSubmission extends IridaResourceSupport implements MutableI
 	 */
 	@JsonIgnore
 	public Set<SequenceFilePair> getPairedInputFiles() {
-		return inputFilesPaired;
+		return inputFiles.stream().filter(f -> {
+			return f instanceof SequenceFilePair;
+		}).map(f -> {
+			return (SequenceFilePair) f;
+		}).collect(Collectors.toSet());
 	}
 
 	/**
@@ -503,8 +507,7 @@ public class AnalysisSubmission extends IridaResourceSupport implements MutableI
 	 */
 	public static class Builder {
 		private String name;
-		private Set<SingleEndSequenceFile> inputFilesSingleEnd;
-		private Set<SequenceFilePair> inputFilesPaired;
+		private Set<SequencingObject> inputFiles;
 		private ReferenceFile referenceFile;
 		private UUID workflowId;
 		private Map<String, String> inputParameters;
@@ -539,31 +542,17 @@ public class AnalysisSubmission extends IridaResourceSupport implements MutableI
 		}
 
 		/**
-		 * Sets this {@link SingleEndSequenceFile}s for this submission
-		 * 
-		 * @param inputFilesSingleEnd
-		 *            {@link SingleEndSequenceFile}s for this submission
-		 * @return a {@link Builder}
-		 */
-		public Builder inputFilesSingleEnd(Set<SingleEndSequenceFile> inputFilesSingleEnd) {
-			checkNotNull(inputFilesSingleEnd);
-			checkArgument(!inputFilesSingleEnd.isEmpty());
-			this.inputFilesSingleEnd = inputFilesSingleEnd;
-			return this;
-		}
-
-		/**
 		 * Sets the inputFilesPaired for this submission.
 		 * 
-		 * @param inputFilesPaired
+		 * @param inputFiles
 		 *            The inputFilesPaired for this submission.
 		 * @return A {@link Builder}.
 		 */
-		public Builder inputFilesPaired(Set<SequenceFilePair> inputFilesPaired) {
-			checkNotNull(inputFilesPaired, "inputFilesPaired is null");
-			checkArgument(!inputFilesPaired.isEmpty(), "inputFilesPaired is empty");
+		public Builder inputFiles(Set<SequencingObject> inputFiles) {
+			checkNotNull(inputFiles, "inputFiles is null");
+			checkArgument(!inputFiles.isEmpty(), "inputFiles is empty");
 
-			this.inputFilesPaired = inputFilesPaired;
+			this.inputFiles = inputFiles;
 			return this;
 		}
 
@@ -649,8 +638,8 @@ public class AnalysisSubmission extends IridaResourceSupport implements MutableI
 		}
 
 		public AnalysisSubmission build() {
-			checkArgument(inputFilesSingleEnd != null || inputFilesPaired != null,
-					"all input file collections are null.  You must supply at least one set of input files");
+			checkArgument(inputFiles != null,
+					"input file collection is null.  You must supply at least one set of input files");
 
 			return new AnalysisSubmission(this);
 		}
@@ -700,9 +689,9 @@ public class AnalysisSubmission extends IridaResourceSupport implements MutableI
 
 	@Override
 	public int hashCode() {
-		return Objects.hash(name, workflowId, remoteAnalysisId, remoteInputDataId, remoteWorkflowId, inputFilesSingleEnd,
-				inputFilesPaired, createdDate, modifiedDate, analysisState, analysisCleanedState, analysis,
-				referenceFile, namedParameters, submitter);
+		return Objects.hash(name, workflowId, remoteAnalysisId, remoteInputDataId, remoteWorkflowId, inputFiles,
+				createdDate, modifiedDate, analysisState, analysisCleanedState, analysis, referenceFile,
+				namedParameters, submitter);
 	}
 
 	@Override
@@ -713,9 +702,7 @@ public class AnalysisSubmission extends IridaResourceSupport implements MutableI
 					&& Objects.equals(name, p.name) && Objects.equals(workflowId, p.workflowId)
 					&& Objects.equals(remoteAnalysisId, p.remoteAnalysisId)
 					&& Objects.equals(remoteInputDataId, p.remoteInputDataId)
-					&& Objects.equals(remoteWorkflowId, p.remoteWorkflowId)
-					&& Objects.equals(inputFilesSingleEnd, p.inputFilesSingleEnd)
-					&& Objects.equals(inputFilesPaired, p.inputFilesPaired)
+					&& Objects.equals(remoteWorkflowId, p.remoteWorkflowId) && Objects.equals(inputFiles, p.inputFiles)
 					&& Objects.equals(analysisState, p.analysisState)
 					&& Objects.equals(analysisCleanedState, p.analysisCleanedState)
 					&& Objects.equals(referenceFile, p.referenceFile)